Luotu TMDB API kutsuille controlleri ja sinne testi-kutsu
