+++ app.js
@@ -1,25 +1,12 @@
 const dotenv = require('dotenv').config();
 var express = require('express');
-const https = require('https');
+
+const TMDB_router = require('./controllers/TMDBController');
 
 const App=express();
 //App.use(express.json);
 
-App.get('/', () => {
-    let data = '';
-    const request = https.get('https://api.themoviedb.org/3/movie/11?api_key=' + process.env.API_KEY, (response) => {
-        console.log('https://api.themoviedb.org/3/movie/11?api_key=' + process.env.API_KEY);
-        response.on('data', (chunk) => {
-            data += chunk;
-            console.log(response.data);
-        });
-        response.on('end', () => {
-            console.log("error");
-            console.log(data);
-        });
-        request.end();
-    });
-});
+App.use('/TMDB', TMDB_router);
 
 App.listen(process.env.PORT, function(){
     console.log("listening port " + process.env.PORT);
